pyproject.toml
[tool.poetry]
name = "dipdup-bets"
version = "0.1.0"
description = ""
authors = ["Lev Gorodetskiy <github@droserasprout.space>"]

[tool.poetry.dependencies]
python = "^3.9"
dipdup = {git = "https://github.com/dipdup-net/dipdup-py.git"}
dateutils = "^0.6.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
